+++ kitty/session.py
@@ -11,7 +11,7 @@ from functools import partial
 from gettext import gettext as _
 from typing import TYPE_CHECKING, Any, Optional, Sequence, Union
 
-from .cli_stub import CLIOptions, SaveAsSessionOptions
+from .cli_stub import CLIOptions, GotoSessionOptions, SaveAsSessionOptions
 from .constants import config_dir, unserialize_launch_flag
 from .fast_data_types import get_options
 from .layout.interface import all_layouts
@@ -493,11 +493,15 @@ def close_session_with_confirm(boss: BossType, cmdline: Sequence[str]) -> None:
         do_close(True)
 
 
-def choose_session(boss: BossType) -> None:
+def choose_session(boss: BossType, opts: GotoSessionOptions) -> None:
     all_known_sessions = get_all_known_sessions()
     hmap = {n: len(goto_session_history)-i for i, n in enumerate(goto_session_history)}
-    def skey(name: str) -> tuple[int, str]:
-        return hmap.get(name, len(goto_session_history)), name.lower()
+    if opts.sort_by == 'alphabetical':
+        def skey(name: str) -> tuple[int, str]:
+            return 0, name.lower()
+    else:
+        def skey(name: str) -> tuple[int, str]:
+            return hmap.get(name, len(goto_session_history)), name.lower()
     names = sorted(all_known_sessions, key=skey)
 
     def chosen(name: str | None) -> None:
@@ -507,9 +511,32 @@ def choose_session(boss: BossType) -> None:
         _('Select a session to activate'), ((name, name) for name in names), chosen)
 
 
+def parse_goto_session_cmdline(args: list[str]) -> tuple[GotoSessionOptions, list[str]]:
+    from kitty.cli import cached_parse_cmdline
+    ans = GotoSessionOptions()
+    leftover_args = cached_parse_cmdline(goto_session_options(), args, ans)
+    return ans, leftover_args
+
+
+
+def goto_session_options() -> str:
+    return '''
+--sort-by
+choices=recent,alphabetical
+default=recent
+When interactively choosing sessions from a list, how to sort the list.
+'''
+
+
 def goto_session(boss: BossType, cmdline: Sequence[str]) -> None:
+    try:
+        opts, cmdline = parse_goto_session_cmdline(list(cmdline))
+    except Exception as e:
+        boss.show_error(_('Invalid goto_session command'), _(
+            'The command goto_session {0} is invalid with error: {1}').format(shlex.join(cmdline), e))
+        return
     if not cmdline:
-        choose_session(boss)
+        choose_session(boss, opts)
         return
     path = cmdline[0]
     if len(cmdline) == 1: